requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.4.4
Flask-Login==0.6.2
Werkzeug==2.3.3
PyMySQL==1.0.2
python-dotenv==1.0.0